refactor: disambiguate env proxy from the unifont api proxy in file names
